@@ -1,12 +1,18 @@
 {
   "name": "@ponchojs/eventbus",
   "version": "0.2.1",
-  "main": "dist/index.js",
+  "module": "dist/main.es.js",
   "types": "dist/index.d.ts",
   "author": "Lukas Spatschil",
   "files": [
     "dist"
   ],
+  "exports": {
+    ".": {
+      "import": "./dist/main.es.js",
+      "require": "./dist/main.umd.js"
+    }
+  },
   "engines": {
     "node": ">=10"
   },
@@ -33,12 +39,13 @@
   ],
   "scripts": {
     "prebuild": "rimraf -rf ./dist && tsc --noEmit",
-    "build": "tsc --emitDeclarationOnly && esbuild lib/index.ts --bundle --minify --outdir=dist --target=chrome58,firefox57,safari11,edge16",
+    "build": "vite build",
     "start": "node ./dist/index.js",
     "test": "jest",
     "test:watch": "jest --watch",
     "test:coverage": "jest --coverage",
-    "lint": "tsc --noEmit"
+    "lint": "tsc --noEmit && eslint . --ext .ts",
+    "semantic-release": "semantic-release"
   },
   "prettier": {
     "printWidth": 160,
@@ -48,15 +55,43 @@
     "trailingComma": "es5"
   },
   "devDependencies": {
-    "@types/jest": "^26.0.23",
-    "@types/node": "^15.12.5",
-    "esbuild": "^0.12.28",
-    "jest": "^27.0.6",
+    "@commitlint/cli": "^17.0.2",
+    "@commitlint/config-conventional": "^17.0.2",
+    "@types/jest": "^28.1.1",
+    "@types/node": "^17.0.42",
+    "@typescript-eslint/eslint-plugin": "^5.28.0",
+    "@typescript-eslint/parser": "^5.28.0",
+    "eslint": "^8.17.0",
+    "husky": "^8.0.1",
+    "jest": "^28.1.1",
     "rimraf": "^3.0.2",
-    "ts-jest": "^27.0.3",
-    "typescript": "^4.3.4"
+    "semantic-release": "^19.0.3",
+    "ts-jest": "^28.0.5",
+    "typescript": "^4.7.3",
+    "vite": "^2.9.12",
+    "vite-plugin-dts": "^1.2.0"
   },
   "jest": {
     "preset": "ts-jest"
+  },
+  "husky": {
+    "hooks": {
+      "commit-msg": "commitlint -E  HUSKY_GIT_PARAMS"
+    }
+  },
+  "commitlint": {
+    "extends": [
+      "@commitlint/config-conventional"
+    ],
+    "rules": {
+      "subject-case": [
+        2,
+        "never",
+        [
+          "start-case",
+          "pascal-case"
+        ]
+      ]
+    }
   }
 }
